CrewAI/requirements.txt
python-dotenv
crewai[tools]
langchain
langsmith
google-search-results
duckduckgo-search
openai
twilio
sendgrid